third_party/blink/web_tests/platform/generic/fast/table/table-sections-border-spacing-expected.txt

Test for chromium bug : 29502. border-spacing is doubled between table-row-groups.
Border-spacing was adding twice, One for bottom of first section and another for top of second section. But bottom of first section and top of second section border spacing should be common.
Distance between all rows in the table should be same.


row0 head0	row0 head1
row0 col0	row0 col11
row1 col0	row1 col11
row0 col0	row0 col11
row1 col0	row1 col11
row0 foot0	row0 foot1




PASS 10 is 10


PASS 10 is 10


PASS 10 is 10


PASS 10 is 10


PASS 10 is 10
PASS successfullyParsed is true

TEST COMPLETE
